requirement.txt
langchain
langchain-community
langchain-chroma
langchain-text-splitters
pyyaml
dashscope
pypdf
